pyproject.toml
[tool.poetry]
name = "robotic-fleet-intelligence-platform"
version = "0.1.0"
description = "Projet Trafic Aérien"
authors = ["IshemZ <ishem.zerzour@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
numpy = "^2.2.5"
scikit-learn = "^1.6.1"
streamlit = "^1.45.0"
requests = "^2.32.3"
pylance = "^0.26.1"
python-dotenv = "^1.1.0"
setuptools = "^80.4.0"
pymongo = "^4.12.1"
watchdog = "^6.0.0"
plotly = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
